requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
boto3==1.24.57
botocore==1.27.57
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==2.4.2
charset-normalizer==2.1.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.3
django-appconf==1.0.5
django-cors-headers==3.13.0
django-imagekit==4.1.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.3
Faker==13.15.1
Flask==2.0.3
fuzzywuzzy==0.18.0
hiredis==2.1.0
httpie==3.2.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
Levenshtein==0.20.5
MarkupSafe==2.1.1
msgpack==0.6.2
multidict==6.0.2
mysqlclient==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pilkit==2.0
Pillow==9.2.0
playsound==1.3.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.6
rapidfuzz==2.11.1
razorpay==1.3.0
redis==4.3.5
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sendotp==0.0.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
whitenoise==6.2.0
zope.interface==5.5.2
